added non-physicals to ViewAnimals

+++ cuACS/CuacsAPI.cc
@@ -26,11 +26,11 @@ vector<Animal*> CuacsAPI::getAnimals()
 
 void CuacsAPI::addAnimal(string name, string animalType, string breed, int age, string gender, string color, string size,
                          int aggression, int attachment, int obedience, string energy, bool isCrateTrained, bool isHypoallergenic,
-                         bool isNeutered, int childrenComfort, int loudness, int cost, int costPerYear, int intelligence, int cleanliness)
+                         bool isNeutered, int childrenComfort, int loudness, float cost, float costPerYear, int intelligence, int cleanliness)
 {
     availableId = availableId + 1;
     ps.storeProfile(new Animal(availableId, name, animalType, breed, age, gender, color, size, aggression, attachment, obedience, energy,
-                               isCrateTrained, isHypoallergenic, isNeutered, childrenComfort, loudness, cost, costPerYear, intelligence, cleanliness ), "Animal");
+                               isCrateTrained, isHypoallergenic, isNeutered, childrenComfort, loudness, cost, costPerYear, intelligence, cleanliness), "Animal");
 }
 
 void CuacsAPI::addHuman(string name, int age)
