+++ cuACS/CuacsAPI.h
@@ -13,7 +13,7 @@ class CuacsAPI
 		vector<Animal*> getAnimals();
         void addAnimal(string name, string animalType, string breed, int age, string gender, string color, string size,
                        int aggression, int attachment, int obedience, string energy, bool isCrateTrained, bool isHypoallergenic,
-                       bool isNeutered, int childrenComfort, int loudness, int cost, int costPerYear, int intelligence, int cleanliness);
+                       bool isNeutered, int childrenComfort, int loudness, float cost, float costPerYear, int intelligence, int cleanliness);
         void addHuman(string name, int age);
 		~CuacsAPI();
